requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
django==3.2.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
psycopg2
gunicorn==20.1.0
django-cors-headers==3.10.0
django-heroku==0.3.1
#psycopg2-binary==2.9.1
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.2
whitenoise==5.3.0
Pillow==7.2.0
